requirements.txt
pbr!=2.1.0,>=2.0.0 # Apache-2.0
jsonschema
referencing
requests
PyYAML>=3.10.0  # MIT
openstackdocstheme>=2.2.1 # Apache-2.0
sphinx>=2.0.0,!=2.1.0 # BSD
